Add email to email tab, merge settings into settings.py file
