+++ app/blueprints/notifications/__init__.py
@@ -17,10 +17,7 @@
 
 from flask import Blueprint, render_template, redirect, url_for
 from flask_login import current_user, login_required
-from flask_wtf import FlaskForm
-from wtforms import BooleanField, SubmitField
-from app.blueprints.users.profile import get_setting_tabs
-from app.models import db, Notification, UserNotificationPreferences, NotificationType
+from app.models import db, Notification
 
 bp = Blueprint("notifications", __name__)
 
@@ -37,45 +34,3 @@ def clear():
 	Notification.query.filter_by(user=current_user).delete()
 	db.session.commit()
 	return redirect(url_for("notifications.list_all"))
-
-
-@bp.route("/notifications/settings/", methods=["GET", "POST"])
-@login_required
-def settings():
-	is_new = False
-	prefs = current_user.notification_preferences
-	if prefs is None:
-		is_new = True
-		prefs = UserNotificationPreferences(current_user)
-
-	attrs = {
-		"submit": SubmitField("Save")
-	}
-
-	data = {}
-	types = []
-	for notificationType in NotificationType:
-		key = "pref_" + notificationType.toName()
-		types.append(notificationType)
-		attrs[key] = BooleanField("")
-		data[key] = getattr(prefs, key) == 2
-
-	SettingsForm = type("SettingsForm", (FlaskForm,), attrs)
-
-	form = SettingsForm(data=data)
-	if form.validate_on_submit():
-		for notificationType in NotificationType:
-			key = "pref_" + notificationType.toName()
-			field = getattr(form, key)
-			value = 2 if field.data else 0
-			setattr(prefs, key, value)
-
-		if is_new:
-			db.session.add(prefs)
-
-		db.session.commit()
-		return redirect(url_for("notifications.settings"))
-
-	return render_template("notifications/settings.html",
-			form=form, user=current_user, types=types, is_new=is_new,
-			tabs=get_setting_tabs(current_user), current_tab="notifications")
