+++ app/blueprints/users/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 bp = Blueprint("users", __name__)
 
-from . import profile, claim, account
+from . import profile, claim, account, settings
